+++ CHANGELOG.md
@@ -2,14 +2,15 @@
 
 ## 1.2.0 - 2019-03-17
 
-- `values()` method to return just the values without the keys
+* `values()` method to return just the values without the keys
+* `keys()` method to return the constant names as an array
 
 ## 1.1.0 - 2018-08-07
 
 ### Added
 
-- `has()` method to check for existence of value and return a boolean
+* `has()` method to check for existence of value and return a boolean
 
 ## 1.0.0 - 2017-12-11
 
-- initial release
+* initial release